refactor: UserDTO 디렉토리 변경
UserDTO의 디렉토리가 domain/auth -> domain/user로 변경되었습니다.
