+++ server/src/domain/user/types/user-dto.ts
